Fixing search dropdown behavior
